+++ app/graphql/types/query_type.rb
@@ -113,7 +113,7 @@ module Types
       argument :coach_id, ID, required: false
     end
 
-    resolved_field :community_topics, Types::TopicDetailsType.connection_type, null: false do
+    resolved_field :topics, Types::TopicType.connection_type, null: false do
       argument :community_id, ID, required: true
       argument :topic_category_id, ID, required: false
       argument :target_id, ID, required: false
