+++ app/javascript/communities/types/CommunitiesShow__Types.re
@@ -1,4 +1,3 @@
 module Topic = CommunitiesShow__Topic;
 module Topics = CommunitiesShow__Topics;
 module Target = CommunitiesShow__Target;
-module TopicCategory = CommunitiesShow__TopicCategory;
