+++ app/javascript/packs/CommunitiesNewTopicPack.re
@@ -3,7 +3,7 @@ open TopicsShow__Types;
 type props = {
   communityId: string,
   target: option(LinkedTarget.t),
-  topicCategories: array(CommunitiesShow__TopicCategory.t),
+  topicCategories: array(TopicCategory.t),
 };
 
 let decodeProps = json =>
@@ -11,11 +11,7 @@ let decodeProps = json =>
     communityId: json |> field("communityId", string),
     target: json |> optional(field("target", LinkedTarget.decode)),
     topicCategories:
-      json
-      |> field(
-           "topicCategories",
-           array(CommunitiesShow__TopicCategory.decode),
-         ),
+      json |> field("topicCategories", array(TopicCategory.decode)),
   };
 
 let props = DomUtils.parseJSONTag() |> decodeProps;
