+++ app/queries/create_community_mutator.rb
@@ -4,7 +4,6 @@ class CreateCommunityMutator < ApplicationQuery
   property :name, validates: { length: { minimum: 1, maximum: 50 } }
   property :target_linkable
   property :course_ids
-  property :topic_categories
 
   validate :course_must_exist_in_current_school
 
@@ -15,26 +14,17 @@ class CreateCommunityMutator < ApplicationQuery
   end
 
   def create_community
-    community = current_school.communities.create!(
+    current_school.communities.create!(
       name: name,
       target_linkable: target_linkable,
       courses: courses,
     )
-    create_categories(community) if topic_categories.present?
-
-    community
   end
 
   private
 
   def resource_school
     current_school
-  end
-
-  def create_categories(community)
-    topic_categories.each do |category|
-      community.topic_categories.create!(name: category)
-    end
   end
 
   def courses
