+++ db/migrate/20200910172304_create_topic_categories.rb
@@ -5,7 +5,7 @@ class CreateTopicCategories < ActiveRecord::Migration[6.0]
       t.string :name, null: false
     end
 
-    add_reference :topics, :topic_category, index: true
+    add_reference :topics, :topic_category, index: true, foreign_key: true
     add_index :topic_categories, %i[name community_id], unique: true
   end
 end
